declare MoveType enum inside game engine module
