+++ server/src/main/java/com/coincoinche/engine/MoveBidding.java
@@ -207,16 +207,13 @@ public class MoveBidding extends Move implements Comparable<MoveBidding> {
     String specialMoveJsonTemplate = "{\"moveType\":\"%s\", \"special\":\"%s\"}";
 
     if (isCoinche()) {
-      return String.format(
-          specialMoveJsonTemplate, PlayerBadeEvent.MoveType.SPECIAL_BIDDING, Special.COINCHE);
+      return String.format(specialMoveJsonTemplate, MoveType.SPECIAL_BIDDING, Special.COINCHE);
     }
     if (isSurcoinche()) {
-      return String.format(
-          specialMoveJsonTemplate, PlayerBadeEvent.MoveType.SPECIAL_BIDDING, Special.SURCOINCHE);
+      return String.format(specialMoveJsonTemplate, MoveType.SPECIAL_BIDDING, Special.SURCOINCHE);
     }
     if (isPass()) {
-      return String.format(
-          specialMoveJsonTemplate, PlayerBadeEvent.MoveType.SPECIAL_BIDDING, Special.PASS);
+      return String.format(specialMoveJsonTemplate, MoveType.SPECIAL_BIDDING, Special.PASS);
     }
 
     return getContract().toJson();
